Converted menu into a pure virtual class
