+++ CMakeLists.txt
@@ -21,7 +21,7 @@ set(SOURCE_FILES
     ${CMAKE_CURRENT_SOURCE_DIR}/src/main.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/nspi_console.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/nspi_pad.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/src/nspi_menu.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/nspi_title_selector_menu.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/nspi_app.cpp
 )
